+++ packages/components-js/tests/vrt/playwright/specs/select-wrapper.vrt.ts
@@ -1,7 +1,7 @@
 import { executeVisualRegressionTest, selectNode } from '../helpers/playwright-helper';
 import { test } from '@playwright/test';
 
-test.skip('should have no visual regression', async () => {
+test.describe('should have no visual regression', async () => {
   await executeVisualRegressionTest('select-wrapper', {
     scenario: async (page) => {
       const btn = await selectNode(