+++ packages/components-js/tests/vrt/playwright/specs/toast.vrt.ts
@@ -1,22 +1,22 @@
 import { executeVisualRegressionTest } from '../helpers/playwright-helper';
 import { test } from '@playwright/test';
 
-test.skip('should have no visual regression', async () => {
+test.describe('should have no visual regression', async () => {
   await executeVisualRegressionTest('toast-basic');
 });
 
-test.skip('should have no visual regression when dark', async () => {
+test.describe('should have no visual regression when dark', async () => {
   await executeVisualRegressionTest('toast-basic-dark');
 });
 
-test.skip('should have no visual regression with long-text', async () => {
+test.describe('should have no visual regression with long-text', async () => {
   await executeVisualRegressionTest('toast-basic-long-text');
 });
 
-test.skip('should have no visual regression with offset', async () => {
+test.describe('should have no visual regression with offset', async () => {
   await executeVisualRegressionTest('toast-offset');
 });
 
-test.skip('should have no visual regression with prefix', async () => {
+test.describe('should have no visual regression with prefix', async () => {
   await executeVisualRegressionTest('toast-prefixed');
 });