+++ packages/components-js/tests/vrt/puppeteer/specs/component-scaling.vrt.ts
@@ -1,7 +1,30 @@
-import { openPopoversAndHighlightSpacer, vrtTest } from '@porsche-design-system/shared/testing';
-import { getVisualRegressionStatesTester } from 'shared/src/testing/vrt';
+import {
+  getVisualRegressionStatesTester,
+  openPopoversAndHighlightSpacer,
+  vrtTest,
+} from '@porsche-design-system/shared/testing';
 
-const components = ['checkbox-wrapper', 'radio-button-wrapper', 'link', 'button', 'tag-dismissible', 'spinner', 'popover', 'segmented-control'];
+const components = [
+  'banner',
+  'button',
+  'button-pure',
+  'checkbox-wrapper',
+  'inline-notification',
+  'link',
+  'link-pure',
+  'link-social',
+  'pagination',
+  'popover',
+  'radio-button-wrapper',
+  'select-wrapper',
+  'spinner',
+  'tag',
+  'tag-dismissible',
+  'text-field-wrapper',
+  'textarea-wrapper',
+  'toast-basic',
+ 'segmented-control'
+];
 
 it.each(components)('should have no visual regression for scaled component %s', async (component) => {
   expect(