+++ packages/components-react/src/pages/ToastBasic.tsx
@@ -15,7 +15,7 @@ export const ToastBasicPage = (): JSX.Element => {
     <>
       <style dangerouslySetInnerHTML={{ __html: style }} />
 
-      <div className="playground light" title="should render toast neutral on light background">
+      <div className="playground light" title="should render toast info on light background">
         <Toast text="Some message" />
       </div>
     </>