+++ packages/components/src/components/textarea-wrapper/textarea-wrapper-styles.ts
@@ -1,49 +1,41 @@
 import type { Styles } from 'jss';
 import type { BreakpointCustomizable, Theme } from '../../types';
-import { buildSlottedStyles, getCss, isVisibleFormState, mergeDeep } from '../../utils';
-import {
-  addImportantToEachRule,
-  getBaseSlottedStyles,
-  pxToRemWithUnit,
-  getThemedColors,
-  getScreenReaderOnlyJssStyle,
-} from '../../styles';
+import { getCss, mergeDeep } from '../../utils';
+import { addImportantToEachRule, getThemedColors, getScreenReaderOnlyJssStyle } from '../../styles';
 import { getBaseChildStyles, getLabelStyles } from '../../styles/form-styles';
 import { getFunctionalComponentRequiredStyles } from '../common/required/required-styles';
 import { getFunctionalComponentStateMessageStyles } from '../common/state-message/state-message-styles';
 import type { FormState } from '../../utils/form/form-state';
-import { textSmallStyle } from '@porsche-design-system/utilities-v2';
+import { spacingStaticMedium, spacingStaticLarge, textSmallStyle } from '@porsche-design-system/utilities-v2';
+import { hostHiddenStyles } from '../../styles/host-hidden-styles';
 
 export const getComponentCss = (
   isDisabled: boolean,
   hideLabel: BreakpointCustomizable<boolean>,
   state: FormState,
-  isCounterVisible: boolean,
-  hasCounter: boolean
+  hasCounter: boolean,
+  theme: Theme
 ): string => {
-  const theme: Theme = 'light';
-  const hasVisibleState = isVisibleFormState(state);
   const { contrastMediumColor } = getThemedColors(theme);
-  const defaultPadding = pxToRemWithUnit(hasVisibleState ? 10 : 11);
 
   return getCss({
     '@global': {
-      ':host': {
+      ':host': addImportantToEachRule({
         display: 'block',
-      },
+        ...hostHiddenStyles,
+      }),
       ...mergeDeep(
         addImportantToEachRule(
           getBaseChildStyles('textarea', state, theme, {
-            // 36 = 2 * 6 + 24 where 6 is the bottom distance and 24 the height of the text
-            padding: isCounterVisible
-              ? [defaultPadding, defaultPadding, pxToRemWithUnit(36)].join(' ')
-              : defaultPadding,
-            resize: 'vertical',
+            font: textSmallStyle.font, // to override line-height
+            padding: hasCounter ? `12px ${spacingStaticMedium} ${spacingStaticLarge}` : `12px ${spacingStaticMedium}`,
           })
         ),
         {
           '::slotted(textarea)': {
-            minHeight: pxToRemWithUnit(192), // min-height should be overridable
+            height: 'auto', // removes !important from getBaseChildStyles
+            minHeight: '200px', // min-height should be overridable
+            resize: 'vertical', // overridable, too
           },
         } as Styles
       ),
@@ -54,18 +46,18 @@ export const getComponentCss = (
       hideLabel,
       state,
       theme,
-      isCounterVisible && {
+      hasCounter && {
         counter: {
           position: 'absolute',
-          bottom: pxToRemWithUnit(6),
-          right: pxToRemWithUnit(12),
+          bottom: '6px',
+          right: '12px',
           zIndex: 1,
-          ...textSmallStyle,
+          font: textSmallStyle.font,
           color: contrastMediumColor,
         },
       }
     ),
-    ...getFunctionalComponentRequiredStyles(theme),
+    ...getFunctionalComponentRequiredStyles(),
     ...getFunctionalComponentStateMessageStyles(theme, state),
     ...(hasCounter && {
       'sr-only': {
@@ -74,8 +66,4 @@ export const getComponentCss = (
       },
     }),
   });
-};
-
-export const getSlottedCss = (host: HTMLElement): string => {
-  return getCss(buildSlottedStyles(host, getBaseSlottedStyles()));
 };