+++ packages/components/src/utils/log/index.ts
@@ -1,0 +1,2 @@
+export * from './warnIfDeprecatedComponentIsUsed';
+export * from './warnIfDeprecatedPropValueIsUsed';